refactor: 상품 상세 오버레이의 재오픈 저장 로직을 훅으로 분리
ProductDetailOverlay가 직접 갖고 있던 '로그인 페이지로 이탈하면 재오픈 정보를 저장하고 닫는' effect를 useSaveReopenProductOnLoginExit 훅으로 옮긴다.

- 오버레이는 저장할 데이터(getReopenData)와 닫기(unmount)만 넘기고, 경로 감시·저장·닫기
  타이밍은 훅이 담당 (컴포넌트에서 라우팅·sessionStorage·수명 지식 제거)
- pathname이 LOGIN으로 바뀌는 기존 트리거를 그대로 유지해, 세션 만료 강제 리다이렉트도 계속 커버
